# Dashboard: Use Images for Explore Template Grid View