--- a/packages/dashboard/src/app/views/exploreTemplates/karma/exploreTemplates.karma.js
+++ b/packages/dashboard/src/app/views/exploreTemplates/karma/exploreTemplates.karma.js
@@ -76,10 +76,10 @@ describe('Grid view', () => {
       : Promise.reject(new Error('could not focus on grid'));
   }
 
-  it('should render', () => {
+  it('should pass accessibility tests', async () => {
     const viewTemplates = fixture.screen.queryByText('Viewing all templates');
-
     expect(viewTemplates).toBeTruthy();
+    await expectAsync(viewTemplates).toHaveNoViolations();
   });
 
   it('should navigate to My Stories', async () => {