--- a/packages/dashboard/src/app/views/myStories/content/storyGridItem/components.js
+++ b/packages/dashboard/src/app/views/myStories/content/storyGridItem/components.js
@@ -39,50 +39,11 @@ CustomCardGridItem.propTypes = {
   $posterHeight: PropTypes.number.isRequired,
 };
 
-// Keeps grid item position contained
-export const Container = styled.div`
-  position: relative;
-`;
-
 // Allows grid item contents to overlay in a set space
 export const CardWrapper = styled.div`
   position: absolute;
   height: 100%;
   width: 100%;
-`;
-
-// Displays story poster if available for story grid item as background of card area
-export const Poster = styled.div`
-  height: 100%;
-  width: 100%;
-  object-fit: fill;
-  border-radius: ${({ theme }) => theme.borders.radius.medium};
-  background: ${({ theme }) => theme.colors.gradient.placeholder};
-`;
-
-// Gradient overlays the story poster
-export const Gradient = styled.div`
-  position: absolute;
-  bottom: 0;
-  height: 67%;
-  width: 100%;
-  border-radius: ${({ theme }) => theme.borders.radius.medium};
-  background: ${({ theme }) => theme.colors.gradient.posterOverlay};
-`;
-
-// Holds story content
-export const Scrim = styled.div`
-  position: absolute;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  padding: 8px;
-  border-radius: ${({ theme }) => theme.borders.radius.medium};
-  background: ${({ theme }) => theme.colors.opacity.black3};
 `;
 
 // Components that create the content to display on top of a poster