… packages/dashboard/src/app/views/exploreTemplates/content/test/content.js
… packages/dashboard/src/app/views/savedTemplates/test/savedTemplates.js
… packages/dashboard/src/app/views/shared/grid/components.js
… packages/dashboard/src/app/views/shared/index.js
… packages/dashboard/src/app/views/shared/savedTemplateGridView.js
… packages/dashboard/src/app/views/shared/stories/templateGridView.js
… packages/dashboard/src/app/views/shared/templateGridView.js
… packages/dashboard/src/app/views/templateDetails/content/relatedGrid/index.js
… packages/dashboard/src/app/views/templateDetails/karma/templateDetails.karma.js
… packages/dashboard/src/components/cardGrid/index.js
… packages/dashboard/src/components/cardGrid/stories/index.js
… packages/dashboard/src/components/cardGridItem/cardPreview.js
… packages/dashboard/src/components/cardGridItem/cardTitle.js
… packages/dashboard/src/components/cardGridItem/index.js
… packages/dashboard/src/components/cardGridItem/stories/index.js
… packages/dashboard/src/components/cardGridItem/test/cardTitle.js
… packages/dashboard/src/components/cardGridItem/types.js
… packages/dashboard/src/components/index.js
… packages/dashboard/src/dataUtils/formattedTemplatesArray.js
… packages/templates/README.md
… packages/templates/src/cli.js
… packages/templates/src/raw/indoor-garden-oasis/metaData.js
… packages/templates/src/raw/plant-based-dyes/metaData.js