## Context

Loading all of the templates as stories in the dashboard is getting a _bit_ slow. This update is to improve performance ✨ 

## Summary

Takes the screenshots generated in #8692  and uses them in the explore templates view 

<img width="1196" alt="Screen Shot 2021-08-20 at 3 39 12 PM" src="https://user-images.githubusercontent.com/10720454/130299901-3e6810ec-449c-4815-8e8c-c01b42bcfc01.png">

## Perf Wins 
|  | Initial Load time | Vitals |
| --- | --- | --- |
| Before | 7918ms | ![Screen Shot 2021-08-20 at 3 27 12 PM](https://user-images.githubusercontent.com/10720454/130299298-e383ec06-8eda-429e-8b06-3ddfd0b429ee.png) |
| After | 2290ms | ![Screen Shot 2021-08-20 at 3 30 38 PM](https://user-images.githubusercontent.com/10720454/130299413-63cb7e8b-6450-4831-bed4-0be1d1609ba8.png) |

Additionally, cuts the render time for the template grid when it receives the templates from 1821.7ms to 287.8ms

This is all measured locally on my macbook pro on high speed internet, but the difference there is awesome!



## Relevant Technical Choices

- Batted a chunk of clean up in here given the span of what was involved to update templates to use an image instead of story structure. 
- Moved the "exploreTemplatesGrid" out of "Shared" views since it's really not
- Moved the grid item styles that are specific to views to shared from my stories update a few weeks ago 
- Clean up peripheral of storybook data and some flakey karma tests
- Use `picture` + `source` to get image fallbacks working - webp first then png. 

## To-do

- There's a few templates that I goofed the slugs on, those are getting fixed here: #8728  - for now, this is a good way of seeing the page still work when the images aren't present. 
- Follow this up with updates to template detail view - it's still using the "story" structure. Cutting scope on this where I can. 
- There's probably more perf wins here on the template view once we finish cleaning out the detail view - could probably decrease 1 render of the template content and do more to decrease return renders. 

## User-facing changes

Explore templates grid now uses images not rendering stories! This means: The grid looks like the one in my stories and when you hover or focus you don't get animations. That's it. 

## Testing Instructions

This is all regression testing, really. Can you: 

- Create a story from a template 
- Do templates in explore templates grid all show an image of their first page? 
- Does grid load LTR and RTL? 
- Can you click through to template details still? 
- Can you use the template details "related templates" grid to click through to explore those details or use that template as a story as well? 


<!-- ignore-task-list-start -->
- [ ] This is a non-user-facing change and requires no QA
<!-- ignore-task-list-end -->

This PR can be tested by following these steps:

1.


## Reviews

### Does this PR have a security-related impact?

No

### Does this PR change what data or activity we track or use?

No

### Does this PR have a legal-related impact?

No

## Checklist

<!-- Check these after PR creation -->

- [x] This PR addresses an existing issue and I have linked this PR to it in ZenHub
- [x] I have tested this code to the best of my abilities
- [x] I have verified accessibility to the best of my abilities ([docs](https://github.com/google/web-stories-wp/blob/main/docs/accessibility-testing.md))
- [x] I have verified i18n and l10n (translation, right-to-left layout) to the best of my abilities
- [x] This code is covered by automated tests (unit, integration, and/or e2e) to verify it works as intended ([docs](https://github.com/google/web-stories-wp/tree/main/docs#testing))
- [x] I have added documentation where necessary
- [x] I have added a matching `Type: XYZ` label to the PR

---

<!--
Please reference the issue(s) this PR addresses.
No URLs, just the issue numbers.
Use "Fixes #123" if it fixes an issue.

NOTE: One reference per line!

Example:

Fixes #123
Partially addresses #456
See #789
-->

Partially addresses #8373