@BrittanyIRL 2021-08-24
> Are these just from those slugs that are getting updated in the separate PR? or something else?

Yup! I specified that in the `to do` section. Those should get fixed with your PR that touches the template data.