@csossi 2021-08-24
Getting some broken images when in Explore Templates:
![image.png](https://images.zenhubusercontent.com/5ea085e4704e72c0a99c6997/56d8dbe7-0f1a-4d65-b26a-2fddc003b496)
![image.png](https://images.zenhubusercontent.com/5ea085e4704e72c0a99c6997/a484a13c-c4f4-49b2-86db-e090fc418512)
![image.png](https://images.zenhubusercontent.com/5ea085e4704e72c0a99c6997/25df12bb-ebcb-4661-8a4d-c821661a6e47)
![image.png](https://images.zenhubusercontent.com/5ea085e4704e72c0a99c6997/aeecdc64-9e69-4e1a-8450-b0bd7ef2e1a6)