@csossi 2021-08-24
Verified in QA - images on first page of template in grid in another ticket