@BrittanyIRL 2021-08-24
> @BrittanyIRL FYI I fixed two template slugs that were still incorrect. And the "should navigate to Explore Templates -> Grid view" karma test seems to be failing

thanks! yeah i'm gonna circle back to this PR shortly, wrapping up something else.